### 2023-05-21 #5481 @editorialbot
[PRE REVIEW]: pymgpipe: microbiome metabolic modeling in Python
**Submitting author:** @ym2877 (<a href="http://orcid.org/0009-0003-4597-3340">Yoli Meydan</a>)
**Repository:** https://github.com/korem-lab/pymgpipe
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.15.0
