+++ paper.bib
@@ -5,7 +5,8 @@
   volume={71},
   pages={509--525},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s13105-015-0390-3}
 }
 
 @article{bar2020reference,
@@ -16,7 +17,8 @@
   number={7836},
   pages={135--140},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2896-2}
 }
 
 @article{mallick2019predictive,
@@ -27,7 +29,8 @@
   number={1},
   pages={3136},
   year={2019},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-019-10927-1}
 }
 
 @article{baldini2019microbiome,
@@ -38,7 +41,8 @@
   number={13},
   pages={2332--2334},
   year={2019},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1101/318485}
 }
 
 @article{diener2020micom,
@@ -49,7 +53,8 @@
   number={1},
   pages={e00606--19},
   year={2020},
-  publisher={Am Soc Microbiol}
+  publisher={Am Soc Microbiol},
+  doi={10.1101/361907}
 }
 
 @article{noecker2022mimosa2,
@@ -60,7 +65,8 @@
   number={6},
   pages={1615--1623},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btac003}
 }
 
 @article{orth2010flux,
@@ -71,7 +77,8 @@
   number={3},
   pages={245--248},
   year={2010},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/nbt.1614}
 }
 
 @article{thiele2010protocol,
@@ -82,7 +89,8 @@
   number={1},
   pages={93--121},
   year={2010},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/nprot.2009.203}
 }
 
 @article{heinken2023genome,
@@ -91,7 +99,8 @@
   journal={Nature Biotechnology},
   pages={1--12},
   year={2023},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41587-022-01628-0}
 }
 
 @article{norsigian2020bigg,
@@ -102,7 +111,8 @@
   number={D1},
   pages={D402--D406},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkz1054}
 }
 
 @article{machado2018fast,
@@ -113,7 +123,8 @@
   number={15},
   pages={7542--7553},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gky537}
 }
 
 @article{kindschuh2023preterm,
@@ -122,7 +133,8 @@
   journal={Nature Microbiology},
   pages={1--14},
   year={2023},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1101/2021.06.14.448190}
 }
 
 @article{heinken2019systematic,
@@ -132,7 +144,8 @@
   volume={7},
   pages={1--18},
   year={2019},
-  publisher={Springer}
[…]
