+++ README.md
@@ -1,4 +1,4 @@
-## [pymgpipe](https://korem-lab.github.io/pymgpipe/) | [![DOI](https://zenodo.org/badge/525586335.svg)](https://zenodo.org/badge/latestdoi/525586335) [![Python package](https://github.com/korem-lab/pymgpipe/actions/workflows/python-package.yml/badge.svg?branch=main)](https://github.com/korem-lab/pymgpipe/actions/workflows/python-package.yml) [![Tests](https://github.com/korem-lab/pymgpipe/actions/workflows/tests.yml/badge.svg?branch=main)](https://github.com/korem-lab/pymgpipe/actions/workflows/tests.yml) [![Docs](https://github.com/korem-lab/pymgpipe/actions/workflows/docs.yml/badge.svg)](https://github.com/korem-lab/pymgpipe/actions/workflows/docs.yml)  
+## [pymgpipe](https://korem-lab.github.io/pymgpipe/) | [![DOI](https://zenodo.org/badge/525586335.svg)](https://zenodo.org/badge/latestdoi/525586335) [![status](https://joss.theoj.org/papers/3f284f37987438428fd09ad3d5bd4871/status.svg)](https://joss.theoj.org/papers/3f284f37987438428fd09ad3d5bd4871) [![Python package](https://github.com/korem-lab/pymgpipe/actions/workflows/python-package.yml/badge.svg?branch=main)](https://github.com/korem-lab/pymgpipe/actions/workflows/python-package.yml) [![Tests](https://github.com/korem-lab/pymgpipe/actions/workflows/tests.yml/badge.svg?branch=main)](https://github.com/korem-lab/pymgpipe/actions/workflows/tests.yml) [![Docs](https://github.com/korem-lab/pymgpipe/actions/workflows/docs.yml/badge.svg)](https://github.com/korem-lab/pymgpipe/actions/workflows/docs.yml)  
 <!-- Pytest Coverage Comment:Begin -->
[…]
