### 2023-06-01 #5481 @fboehm
@amoeba @nicoguaro @rougier - would you be willing to review this submission for JOSS? I'd really appreciate reviews from someone with your python expertise. Thank you for considering this invitation.
