### 2023-06-01 #5481 @fboehm
Thanks for getting back to me, @amoeba. And thanks for your continued involvement with JOSS. i hope that we can work together in the future.
