### 2023-06-10 #5481 @fboehm
@mbhall88 @ammaraziz - would you be willing to review this submission for JOSS? Our reviews differ from those of many journals in that we use pre-specified checklists as the focus of the review. The review is also iterative and open, where we encourage communications between the reviewers and authors. Thank you for considering this invitation!
