### 2023-06-11 #5481 @flekschas
>@cumbof @ManavalanG @flekschas - would you be willing to review this submission for JOSS? Our reviews differ from those of many journals in that we use pre-specified checklists as the focus of the review. The review is also iterative and open, where we encourage communications between the reviewers and authors. Thank you for considering this invitation!

@fboehm I don't have any expertise in microbiome research or metabolomic modeling so I don't think I'm suited to review this submission
