### 2023-06-12 #5481 @fboehm
Thanks so much, @ammaraziz ! Yes, it's ok that you don't have experience with modeling microbiomes. If you find that you can't evaluate some aspects of the software, please let me know. I'm confident that your python skills will be quite useful for the review.
