+++ .github/workflows/python-package.yml
@@ -80,5 +80,5 @@ jobs:
     - name: Update version
       uses: brettdorrans/write-version-to-file@master
       with:
-        filename: '/.VERSION'
+        filename: '.VERSION'
         placeholder: '${VERSION}'
