+++ .github/workflows/python-package.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   build:
-    if: ${{ github.event.head_commit.message != 'Github Actions- Push'}}
+    if: ${{ github.event.head_commit.message != 'Github Actions- Update README'}}
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -54,14 +54,21 @@ jobs:
       if: ${{ github.ref == 'refs/heads/main' }}
       run: |
         sed -i '/<!-- Pytest Coverage Comment:Begin -->/,/<!-- Pytest Coverage Comment:End -->/c\<!-- Pytest Coverage Comment:Begin -->\n\${{ steps.coverageComment.outputs.coverageHtml }}\n<!-- Pytest Coverage Comment:End -->' ./README.md
+    - name: Commit & Push changes to Readme
+      if: ${{ github.ref == 'refs/heads/main' }} 
+      uses: actions-js/push@master
+      with:
+        github_token: ${{ secrets.PUSH_README }}
+        message: 'Github Actions- Update README'
   release:
     needs: build
-    if: ${{ github.ref == 'refs/heads/main' }}  && ${{ github.event.head_commit.message != 'Github Actions- Push'}}
+    if: ${{ github.ref == 'refs/heads/main' }}  && ${{ github.event.head_commit.message != 'Github Actions- Update README'}}
     runs-on: ubuntu-latest
     env:
       GITHUB_TOKEN: ${{ secrets.RELEASE }}
     steps:
-      - uses: rymndhng/release-on-push-action@master
+      - id: release
+        uses: rymndhng/release-on-push-action@master
         with:
           tag_prefix: "v"
           bump_version_scheme: norelease
@@ -74,16 +81,11 @@ jobs:
     - name: Update version
       uses: brettdorrans/write-version-to-file@master
       with:
-        filename: '.VERSION'
+        filename: '/.VERSION'
         placeholder: '${VERSION}'
-  push-changes:
-    if: ${{ github.ref == 'refs/heads/main' }}  && ${{ github.event.head_commit.message != 'Github Actions- Push'}}
-    runs-on: ubuntu-latest
-    name: Write Version to File
-    steps:
-    - name: Commit & Push Changes
+    - name: Commit & Push changes to Readme
       if: ${{ github.ref == 'refs/heads/main' }} 
       uses: actions-js/push@master
       with:
         github_token: ${{ secrets.PUSH_README }}
-        message: 'Github Actions- Push'
+        message: 'Github Actions- Update README'
