+++ .github/workflows/python-package.yml
@@ -54,6 +54,11 @@ jobs:
       if: ${{ github.ref == 'refs/heads/main' }}
       run: |
         sed -i '/<!-- Pytest Coverage Comment:Begin -->/,/<!-- Pytest Coverage Comment:End -->/c\<!-- Pytest Coverage Comment:Begin -->\n\${{ steps.coverageComment.outputs.coverageHtml }}\n<!-- Pytest Coverage Comment:End -->' ./README.md
+    - name: Update version
+      uses: brettdorrans/write-version-to-file@master
+      with:
+        filename: '.VERSION'
+        placeholder: '${VERSION}'
     - name: Commit & Push changes to Readme
       if: ${{ github.ref == 'refs/heads/main' }} 
       uses: actions-js/push@master
@@ -73,22 +78,4 @@ jobs:
           tag_prefix: "v"
           bump_version_scheme: norelease
           use_github_release_notes: true
-  write-version:
-    runs-on: ubuntu-latest
-    name: Write Version to File
-    steps:
-    - uses: actions/checkout@v3
-      with:
-        persist-credentials: false 
-        fetch-depth: 0 
-    - name: Update version
-      uses: brettdorrans/write-version-to-file@master
-      with:
-        filename: '.VERSION'
-        placeholder: '${VERSION}'
-    - name: Commit & Push changes to Readme
-      if: ${{ github.ref == 'refs/heads/main' }} 
-      uses: actions-js/push@master
-      with:
-        github_token: ${{ secrets.PUSH_README }}
-        message: 'Github Actions- Update README'
+
