### 2023-07-03 korem-lab/pymgpipe#6 @cumbof
JOSS Review
Hello,

I have a few comments concerning your python package and your manuscript:

1. let's start with the installation process. I honestly don't see a reason why this package shouldn't be available in the Python Package Index [https://pypi.org](https://pypi.org). I would strongly suggest to create a pypi package. You already have everything to do that;
2. before uploading the package to pypi, I would definitely refactor the `setup.py`. Please add as many details as possible (see this example [here](https://github.com/cumbof/hdlib/blob/main/setup.py));
3. this is the first time I see taking track of a tool version through a `.VERSION` file. Is that used for any other purpose? Is it used by a GitHub action? If not, I would add this info in the `pymgpipe/__init__.py`, i.e., `__version__ = "v0.16.0"`. At that point, I would add `from pymgpipe import __version__` in the `setup.py` to access the software version, without the need to read a file. And you can get rid of the `.VERSION` file now;
4. now that you have your package available in pypi, it would be extremely easy to define a Bioconda recipe. At that point, once you have a conda package, you don't have to take care of installing the external dependencies by your own anymore;
5. on line 65-66, you said that `The pymgpipe python package, as well as all associated documentation, tests, and example workflows, can be found at https://korem-lab.github.io/pymgpipe`. This is not correct since that page shows the code documentation only. I would reformulate that sentence;
6. the code documentation seems not complete. Most of the functions are not documented. This must be fixed;
7. that's great that you provide a Notebook to better explain users how to use the library. The problem here is the lack of notes. I can only see python codes here, but there is no explanation about what you are doing. Please, add a text explaining what you do in every cell of the notebook.

Please, do not close this thread.
I'll update it in case I will have any other comments.
