### 2023-07-03 #5545 @cumbof
@fboehm I've updated my checklist.
I left a few points not marked since they refer to my comments in korem-lab/pymgpipe#6
