### 2023-07-03 korem-lab/pymgpipe#6 @ym2877
@cumbof thank you for taking the time to review our package! These points are very helpful, I will try my best to address each of them over the next couple of days. Will keep you posted! :)
