+++ setup.py
@@ -1,25 +1,48 @@
 from setuptools import setup, find_packages
 
-with open('.VERSION') as f:
-    version = f.readline().strip()[1:]
-
 setup(
+    author="Yoli Meydan",
+    author_email="ym2877@cumc.columbia.edu",
     name="pymgpipe",
-    version=version,
+    description="Community level microbiome metabolic modeling in Python",
+    version=open("pymgpipe/resources/.VERSION").readline().strip()[1:],
+    classifiers=[
+        "Intended Audience :: Healthcare Industry",
+        "Intended Audience :: Information Technology",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: Apache Software License",
+        "Programming Language :: Python :: 3",
+        "Topic :: Scientific/Engineering",
+    ],
     install_requires=[
         "cobra",
         "optlang",
+        "scikit_bio",
+        "optlang",
         "tqdm",
+        "seaborn",
     ],
     url="https://github.com/korem-lab/pymgpipe",
     package_dir={"pymgpipe": "pymgpipe"},
-    packages=["pymgpipe"],
+    packages=find_packages(include=["pymgpipe"]),
     package_data={
         "pymgpipe": [
             "resources/diets/*.txt",
             "resources/models/*",
             "resources/problems/*",
             "resources/miniTaxa/*",
+            "resources/.VERSION"
         ],
     },
+    project_urls={
+        "Issues": "https://github.com/korem-lab/pymgpipe/issues",
+        "Source": "https://github.com/korem-lab/pymgpipe",
+        "Readme": "https://github.com/korem-lab/pymgpipe/blob/main/README.md",
+    },
+    python_requires=">=3",
+    license="Apache-2.0",
+    license_files=["LICENSE"],
+    long_description=open("README.md").read(),
+    long_description_content_type="text/markdown",
 )
+
