+++ pymgpipe/__init__.py
@@ -5,3 +5,4 @@ from .nmpc import *
 from .diet import *
 from .coupling import *
 from .modeling import *
+
