+++ .github/workflows/python-package.yml
@@ -78,7 +78,7 @@ jobs:
     - name: Update version
       uses: brettdorrans/write-version-to-file@master
       with:
-        filename: '.VERSION'
+        filename: 'pymgpipe/resources/.VERSION'
         placeholder: '${VERSION}'
     - name: Commit & Push changes to Readme
       if: ${{ github.ref == 'refs/heads/main' }} 
