+++ .github/workflows/python-package.yml
@@ -65,7 +65,7 @@ jobs:
         uses: rymndhng/release-on-push-action@master
         with:
           tag_prefix: "v"
-          bump_version_scheme: norelease
+          bump_version_scheme: minor
           use_github_release_notes: true
   push-changes:
     runs-on: ubuntu-latest
