+++ .github/workflows/python-package.yml
@@ -66,7 +66,7 @@ jobs:
         uses: rymndhng/release-on-push-action@master
         with:
           tag_prefix: "v"
-          bump_version_scheme: minor
+          bump_version_scheme: patch
           use_github_release_notes: true
   push-changes:
     runs-on: ubuntu-latest
