+++ pymgpipe/metrics.py
@@ -1,7 +1,7 @@
 from .utils import get_abundances
 import cobra
 
-def compute_diversity_metrics(model):
+def _compute_diversity_metrics(model):
     assert isinstance(model, cobra.Model), '`model` needs to be COBRA model to compute diversity metrics.'
     print('Computing metrics for %s...'%model.name)
 
