+++ pymgpipe/tests/test_utils.py
@@ -3,7 +3,7 @@ from pymgpipe import *
 def test_remove_reverse_reactions(mini_optlang_model):
     num_reactions = len(mini_optlang_model.variables)
     some_var = mini_optlang_model.variables[100]
-    reverse_var_id = get_reverse_id(some_var.name)
+    reverse_var_id = _get_reverse_id(some_var.name)
     assert reverse_var_id in mini_optlang_model.variables
 
     remove_reverse_vars(mini_optlang_model,hard_remove=False)
