### 2023-07-06 korem-lab/pymgpipe#6 @ym2877
@cumbof I have made the suggested changes and updated the joss branch as well. Please don't hesitate to let me know if you have any additional questions, comments, concerns. Thank you!
