### 2023-07-07 korem-lab/pymgpipe#6 @cumbof
Seems great overall. Thanks for working on my comments @ym2877.

I only have a couple of extra comments:
1. you addressed point 3 by moving the `.VERSION` file into the `pymgpipe/resources` as specified in the `setup.py` file. However, I don't actually see any `.VERSION` file under that folder. I'm ok with taking track of the package version with the `.VERSION` file instead of specifying it in the `pymgpipe/__init__.py` file as I suggested. Just remember to push that file into the resources folder;
2. what about point 4? That's a very minor point compared to the others and it would be enough to have the package in PyPI. However, making it available in bioconda would be way better and it will allow people to avoid taking care of installing the external software dependencies by their own.

Point 1 must be addressed, while point 2 is strongly recommended but not mandatory.
