+++ setup.py
@@ -32,7 +32,6 @@ setup(
             "resources/models/*",
             "resources/problems/*",
             "resources/miniTaxa/*",
-            "resources/.VERSION"
         ],
     },
     project_urls={
