### 2023-07-07 korem-lab/pymgpipe#6 @ym2877
Okay, addressed points #1 and #2! Thanks again for your comments @cumbof. Point #1 was just an error, that reference to `.VERSION` was old and shouldn't have been there, so thank you for catching that!

With regard to #3, I will absolutely look into this. Although, with the way our dependencies are currently set up in `setup.py`, there is no need for users to install external dependencies on their own. The only steps they would need to run is
1. `conda create -n my_env python=3.10`
2. `conda activate my_env`
3. `pip install pymgpipe`

Maybe I am misunderstanding what you mean. Either way, I will look into publishing a bioconda package asap!
