### 2023-07-07 korem-lab/pymgpipe#6 @cumbof
Thanks for your quick reply @ym2877.

That's true, your tool doesn't have any non-python dependencies actually and everything is handled by pip. However, I would consider adding your tool to bioconda in case anyone wants to use your tool as a dependency along with other non-python tools.

Again, I don't consider this point as mandatory in the context of the JOSS review, but I would strongly recommend to consider addressing this extra step. Defining a conda recipe would be extremely easy at this point.

I'm going to close this issue since I don't have any other comments.
