### 2023-07-30 #5545 @fboehm
@ym2877 - A few doi fixes are needed for the references:

- [ ] fix doi for 2019 baldini, et al. I think it needs to be: [10.1093/bioinformatics/bty941](http://dx.doi.org/10.1093/bioinformatics/bty941)
- [ ] fix doi for Diener et al. I think you should use: [10.1128/mSystems.00606-19](http://dx.doi.org/10.1128/mSystems.00606-19)
- [ ] fix Kindschuh et al doi to: [10.1038/s41564-022-01293-8](http://dx.doi.org/10.1038/s41564-022-01293-8)
- [ ] update archive to ensure that the manuscript has the correct dois for references (ie, after making the above changes)

The text of the manuscript looks great - I have no suggestions for it. The references' journal titles, etc., are capitalized as needed. With the above exceptions, the hyperlinks and dois in the references section lead to the correct targets.

Thanks again!
