+++ paper.bib
@@ -42,7 +42,7 @@
   pages={2332--2334},
   year={2019},
   publisher={Oxford University Press},
-  doi={10.1101/318485}
+  doi={10.1093/bioinformatics/bty941}
 }
 
 @article{diener2020micom,
@@ -54,7 +54,7 @@
   pages={e00606--19},
   year={2020},
   publisher={Am Soc Microbiol},
-  doi={10.1101/361907}
+  doi={10.1128/mSystems.00606-19}
 }
 
 @article{noecker2022mimosa2,
@@ -134,7 +134,7 @@
   pages={1--14},
   year={2023},
   publisher={Nature Publishing Group},
-  doi={10.1101/2021.06.14.448190}
+  doi={10.1038/s41564-022-01293-8}
 }
 
 @article{heinken2019systematic,
